ecommerce/requirements.txt
appdirs==1.4.3
asgiref==3.2.3
beautifulsoup4==4.8.1
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
csrf==0.1b1
cssselect2==0.2.2
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
dj-database-url==0.5.0
Django==2.2.5
django-appconf==1.0.3
django-bootstrap4==1.0.1
django-ckeditor==5.7.1
django-crispy-forms==1.7.2
django-import-export==2.0.2
django-js-asset==1.2.2
django-paypal==1.0.0
django-weasyprint==0.5.5
djangorestframework==3.10.3
et-xmlfile==1.0.1
filelock==3.0.12
freeze==1.0.10
html5lib==1.0.1
idna==2.9
importlib-metadata==1.5.0
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.3
path==13.2.0
paypal==1.2.5
Pillow==7.0.0
pipenv==2018.11.26
pycparser==2.20
Pyphen==0.9.5
pytz==2019.3
PyYAML==5.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.0
tablib==1.1.0
tinycss2==1.0.2
urllib3==1.25.8
virtualenv==16.7.5
virtualenv-clone==0.5.3
WeasyPrint==44
webencodings==0.5.1
wfastcgi==3.0.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.0.0
